# terrasignal/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CompOut": {
        "properties": {
          "comp_id": {
            "title": "Comp Id",
            "type": "string"
          },
          "free_rent_months": {
            "title": "Free Rent Months",
            "type": "integer"
          },
          "rent_psf": {
            "title": "Rent Psf",
            "type": "number"
          },
          "signed_date": {
            "title": "Signed Date",
            "type": "string"
          },
          "submarket": {
            "title": "Submarket",
            "type": "string"
          },
          "term_months": {
            "title": "Term Months",
            "type": "integer"
          },
          "ti_allowance_psf": {
            "title": "Ti Allowance Psf",
            "type": "number"
          }
        },
        "required": [
          "comp_id",
          "submarket",
          "signed_date",
          "rent_psf",
          "term_months",
          "ti_allowance_psf",
          "free_rent_months"
        ],
        "title": "CompOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RationaleResponse": {
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "fallback_used": {
            "title": "Fallback Used",
            "type": "boolean"
          },
          "guard_passed": {
            "title": "Guard Passed",
            "type": "boolean"
          },
          "label": {
            "default": "AI-generated explanation",
            "title": "Label",
            "type": "string"
          },
          "memo": {
            "title": "Memo",
            "type": "string"
          },
          "payload_hash": {
            "title": "Payload Hash",
            "type": "string"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          }
        },
        "required": [
          "prediction_id",
          "memo",
          "backend",
          "payload_hash",
          "guard_passed",
          "fallback_used"
        ],
        "title": "RationaleResponse",
        "type": "object"
      },
      "RentForecastRequest": {
        "properties": {
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          }
        },
        "required": [
          "unit_id"
        ],
        "title": "RentForecastRequest",
        "type": "object"
      },
      "RentForecastResponse": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "asset_class": {
            "title": "Asset Class",
            "type": "string"
          },
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          },
          "comp_median_rent_6m": {
            "title": "Comp Median Rent 6M",
            "type": "number"
          },
          "comps": {
            "items": {
              "$ref": "#/components/schemas/CompOut"
            },
            "title": "Comps",
            "type": "array"
          },
          "current_rent_psf": {
            "title": "Current Rent Psf",
            "type": "number"
          },
          "current_tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Tenant Id"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/ShapDriverOut"
            },
            "title": "Drivers",
            "type": "array"
          },
          "lease_expiration": {
            "format": "date",
            "title": "Lease Expiration",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "integer"
          },
          "p10": {
            "title": "P10",
            "type": "number"
          },
          "p50": {
            "title": "P50",
            "type": "number"
          },
          "p90": {
            "title": "P90",
            "type": "number"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "property_name": {
            "title": "Property Name",
            "type": "string"
          },
          "submarket": {
            "title": "Submarket",
            "type": "string"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          },
          "unit_rsf": {
            "title": "Unit Rsf",
            "type": "number"
          }
        },
        "required": [
          "prediction_id",
          "unit_id",
          "property_name",
          "submarket",
          "asset_class",
          "unit_rsf",
          "p10",
          "p50",
          "p90",
          "current_rent_psf",
          "comp_median_rent_6m",
          "lease_expiration",
          "current_tenant_id",
          "as_of",
          "model_version",
          "baseline_mode",
          "drivers",
          "comps"
        ],
        "title": "RentForecastResponse",
        "type": "object"
      },
      "RentQueueItem": {
        "properties": {
          "asset_class": {
            "title": "Asset Class",
            "type": "string"
          },
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          },
          "current_rent_psf": {
            "title": "Current Rent Psf",
            "type": "number"
          },
          "floor": {
            "title": "Floor",
            "type": "integer"
          },
          "lease_expiration": {
            "format": "date",
            "title": "Lease Expiration",
            "type": "string"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "p10": {
            "title": "P10",
            "type": "number"
          },
          "p50": {
            "title": "P50",
            "type": "number"
          },
          "p90": {
            "title": "P90",
            "type": "number"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "property_name": {
            "title": "Property Name",
            "type": "string"
          },
          "submarket": {
            "title": "Submarket",
            "type": "string"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          },
          "unit_rsf": {
            "title": "Unit Rsf",
            "type": "integer"
          },
          "upside_pct": {
            "title": "Upside Pct",
            "type": "number"
          }
        },
        "required": [
          "prediction_id",
          "unit_id",
          "property_name",
          "submarket",
          "market",
          "asset_class",
          "unit_rsf",
          "floor",
          "p50",
          "p10",
          "p90",
          "current_rent_psf",
          "upside_pct",
          "lease_expiration",
          "baseline_mode"
        ],
        "title": "RentQueueItem",
        "type": "object"
      },
...
      "ShapDriverOut": {
        "properties": {
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "shap": {
            "title": "Shap",
            "type": "number"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "feature",
          "label",
          "value",
          "shap"
        ],
        "title": "ShapDriverOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CRE rent forecasting & tenant default-risk platform.",
    "title": "TerraSignal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/forecasts/queue": {
      "get": {
        "operationId": "rent_queue_api_v1_forecasts_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RentQueueItem"
                  },
                  "title": "Response Rent Queue Api V1 Forecasts Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rent Queue",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/v1/forecasts/rent": {
      "post": {
        "operationId": "forecast_rent_api_v1_forecasts_rent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Forecast Rent",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/v1/forecasts/{prediction_id}/rationale": {
      "post": {
        "operationId": "rationale_api_v1_forecasts__prediction_id__rationale_post",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prediction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RationaleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rationale",
        "tags": [
          "forecasts"
        ]
      }
    },
...
  }
}